create concert_info.php
Write basic HTML code and applied it into the main template. The CSS of
the panel is rewritten in this same PHP source file, but not inherent
with the panel in the user_profile.php, in which they share very
similar appearance but my code is cleaner. (It uses too many div
cascades to render this panel.)